refactored navigation-related code
