+++ frontend/src/screens/SignupScreen.tsx
@@ -7,17 +7,17 @@ import CenteredContainer from "../components/CenteredContainer";
 import FormikInput from "../components/FormikInput";
 import LineBreak from "../components/LineBreak";
 import { HomeScreenName } from "./HomeScreen";
-import { useRootScreen } from "./RootScreensManager";
 import { MeDocument, MeQuery, useRegisterMutation } from "../generated/graphql";
 import { useAuthSkip } from "../utils/hooks/useAuthSkip";
 import { useIsFocused } from "@react-navigation/core";
+import { useRootScreen } from "../utils/hooks/useRootScreen";
 
 export const SignupScreenName = "Sign Up";
 
 export type SignupScreenParams = { usernameOrEmail: string | undefined };
 
 function SignupScreen() {
-  const { navigation, route } = useRootScreen(SignupScreenName);
+  const { navigation, route } = useRootScreen();
   useAuthSkip(navigation, useIsFocused());
   const [register] = useRegisterMutation();
