+++ frontend/src/utils/hooks/useAuthSkip.ts
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { useMeQuery } from "../../generated/graphql";
 import { HomeScreenName } from "../../screens/HomeScreen";
-import { RootNavigation } from "../../screens/RootScreensManager";
+import { RootNavigation } from "../types/navigationTypes";
 
 export const useAuthSkip = (navigation: RootNavigation, isFocused: boolean) => {
   const { data, loading } = useMeQuery();
